requirements.txt
wheel==0.36.2
tqdm==4.61.2
six==1.16.0
setuptools==52.0.0.post20210125
ruamel-yaml-conda==0.15.100
requests==2.25.1
pycparser==2.20
pyOpenSSL==20.0.1
pip==21.1.3
idna==2.10
cryptography==3.4.7
conda-package-handling==1.7.3
conda==4.10.3
chardet==4.0.0
cffi==1.14.6
certifi==2021.5.30
brotlipy==0.7.0
PySocks==1.7.1
pycosat==0.6.3
PyYAML==6.0.2
scipy==1.10.1
spikingjelly==0.0.0.0.14
google-auth==2.38.0
tensorboard-data-server==0.7.2
protobuf==5.29.3
grpcio==1.70.0
regex==2024.11.6
google-auth-oauthlib==1.0.0
tensorboard==2.14.0
torchinfo==1.8.0
fsspec==2025.3.0
filelock==3.16.1
huggingface-hub==0.29.3
timm==0.6.12
fastrlock==0.8.3
cupy-cuda113==10.6.0
typing_extensions==4.12.2
smmap==5.0.2
urllib3==1.26.20
pydantic_core==2.27.2
gitdb==4.0.12
annotated-types==0.7.0
setproctitle==1.3.5
sentry-sdk==2.22.0
pydantic==2.10.6
platformdirs==4.3.6
GitPython==3.1.44
eval_type_backport==0.2.2
docker-pycreds==0.4.0
click==8.1.8
wandb==0.19.8
tokenizers==0.20.3
safetensors==0.5.3
transformers==4.46.3
threadpoolctl==3.5.0
joblib==1.4.2
scikit-learn==1.3.2
zipp==3.8.0
widgetsnbextension==3.6.0
websocket-client==1.3.3
webencodings==0.5.1
wcwidth==0.2.5
traitlets==5.3.0
tornado==6.1
torchvision==0.12.0+cu113
torch==1.11.0+cu113
tinycss2==1.1.1
terminado==0.15.0
tensorboard-plugin-wit==1.8.1
supervisor==4.2.4
stack-data==0.3.0
soupsieve==2.3.2.post1
sniffio==1.2.0
rsa==4.8
requests-oauthlib==1.3.1
pyzmq==23.2.0
pytz==2022.1
python-dateutil==2.8.2
pyrsistent==0.18.1
pyparsing==3.0.9
pyasn1-modules==0.2.8
pyasn1==0.4.8
pure-eval==0.2.2
ptyprocess==0.7.0
psutil==5.9.1
prompt-toolkit==3.0.29
prometheus-client==0.14.1
pickleshare==0.7.5
pexpect==4.8.0
parso==0.8.3
pandocfilters==1.5.0
packaging==21.3
oauthlib==3.2.0
numpy==1.22.4
notebook-shim==0.1.0
notebook==6.4.12
nest-asyncio==1.5.5
nbformat==5.4.0
nbconvert==6.5.0
nbclient==0.6.4
nbclassic==0.3.7
mistune==0.8.4
matplotlib-inline==0.1.3
matplotlib==3.5.2
kiwisolver==1.4.3
jupyterlab-widgets==1.1.0
jupyterlab-server==2.14.0
jupyterlab-pygments==0.2.2
jupyterlab-language-pack-zh-CN==3.4.post1
jupyterlab==3.4.3
jupyter-server==1.17.1
jupyter-core==4.10.0
jupyter-client==7.3.4
jsonschema==4.6.0
json5==0.9.8
jedi==0.18.1
ipywidgets==7.7.0
ipython-genutils==0.2.0
ipython==8.4.0
ipykernel==6.15.0
importlib-resources==5.8.0
importlib-metadata==4.11.4
fonttools==4.33.3
fastjsonschema==2.15.3
executing==0.8.3
entrypoints==0.4
defusedxml==0.7.1
decorator==5.1.1
debugpy==1.6.0
cycler==0.11.0
cachetools==5.2.0
bleach==5.0.0
beautifulsoup4==4.11.1
backcall==0.2.0
attrs==21.4.0
asttokens==2.0.5
argon2-cffi-bindings==21.2.0
argon2-cffi==21.3.0
anyio==3.6.1
absl-py==1.1.0
Werkzeug==2.1.2
Send2Trash==1.8.0
Pygments==2.12.0
Pillow==9.1.1
MarkupSafe==2.1.1
Markdown==3.3.7
Jinja2==3.1.2
Babel==2.10.3
